flytt beregning av aft refusjon til egen klasse
